% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Critcorr.R
\name{Critcorr}
\alias{Critcorr}
\title{Cointegration tests}
\usage{
Critcorr(listInput)
}
\arguments{
\item{listInput}{array of price arrays}
}
\description{
This function provides info about linear regression coefs between list components and executes various cointegration tests on it
}
\examples{
Critcorr()
}
\keyword{cointegration}

