% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listAdd.R
\name{listAdd}
\alias{listAdd}
\title{List add}
\usage{
listAdd(listInput)
}
\arguments{
\item{listInput}{array of price arrays}
}
\description{
This function adds 42 to elements of list.
}
\examples{
listAdd()
}
\keyword{cointegration}

