% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regr.R
\name{regr}
\alias{regr}
\title{Cointegration tests}
\usage{
regr(listInput)
}
\arguments{
\item{listInput}{array of price arrays}
}
\description{
This function provides info about linear regression coefs between list components and executes various cointegration tests on it
}
\examples{
regr()
}
\keyword{cointegration}

